refactor: 사이드바 UI 로고/네브 로직 분리
