+++ src/widgets/sidebar/sidebar.types.ts
@@ -4,7 +4,7 @@ export interface TeamProps {
   imageUrl?: string | null;
 }
 
-export interface SideBarUIProps {
+export interface SideBarProps {
   expanded: boolean; // hover 등으로 펼침 여부
   activePath: string; // 현재 경로
   isAlarmOpen: boolean;
